refactor: ComponentCardLoading deleted
