+++ src/library-authoring/components/LibraryComponents.test.tsx
@@ -50,7 +50,7 @@ const blockTypeData = {
   ],
 };
 
-jest.mock('../data/apiHook', () => ({
+jest.mock('../data/apiHooks', () => ({
   useLibraryComponents: () => mockUseLibraryComponents(),
   useLibraryComponentCount: () => mockUseLibraryComponentCount(),
   useLibraryBlockTypes: () => mockUseLibraryBlockTypes(),
@@ -93,11 +93,6 @@ describe('<LibraryComponents />', () => {
     });
     render(<RootWrapper />);
     expect(await screen.findByText(/you have not added any content to this library yet\./i));
-  });
-
-  it('should render loading', async () => {
-    render(<RootWrapper />);
-    expect((await screen.findAllByTestId('card-loading'))[0]).toBeInTheDocument();
   });
 
   it('should render components in full variant', async () => {
@@ -166,22 +161,5 @@ describe('<LibraryComponents />', () => {
     fireEvent.scroll(window, { target: { scrollY: 1000 } });
 
     expect(mockFetchNextPage).not.toHaveBeenCalled();
-  });
-
-  it('should render content and loading when fetching next page', async () => {
-    mockUseLibraryComponents.mockReturnValue({
-      ...data,
-      hits: libraryComponentsMock,
-      isFetching: true,
-      isFetchingNextPage: true,
-      hasNextPage: true,
-    });
-
-    render(<RootWrapper variant="full" />);
-
-    expect(await screen.findByText('This is a text: ID=1')).toBeInTheDocument();
-    expect(screen.getByText('This is a problem: ID=6')).toBeInTheDocument();
-
-    expect((await screen.findAllByTestId('card-loading'))[0]).toBeInTheDocument();
   });
 });
